+++ interchaintest/fee_middleware_test.go
@@ -16,7 +16,7 @@ import (
 	"go.uber.org/zap/zaptest"
 )
 
-func TestScenarioFeeMiddleware(t *testing.T) {
+func TestRelayerFeeMiddleware(t *testing.T) {
 	if testing.Short() {
 		t.Skip()
 	}
@@ -180,7 +180,7 @@ func TestScenarioFeeMiddleware(t *testing.T) {
 	)
 
 	// Wait for relayer to run
-	err = testutil.WaitForBlocks(ctx, 5, chainA, chainB)
+	err = testutil.WaitForBlocks(ctx, 10, chainA, chainB)
 	require.NoError(t, err)
 
 	// Assigning denom