+++ relayer/chains/cosmos/tx.go
@@ -20,6 +20,7 @@ import (
 	tmtypes "github.com/cometbft/cometbft/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/tx"
+	"github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
@@ -37,7 +38,9 @@ import (
 	host "github.com/cosmos/ibc-go/v7/modules/core/24-host"
 	ibcexported "github.com/cosmos/ibc-go/v7/modules/core/exported"
 	tmclient "github.com/cosmos/ibc-go/v7/modules/light-clients/07-tendermint"
+	localhost "github.com/cosmos/ibc-go/v7/modules/light-clients/09-localhost"
 	strideicqtypes "github.com/cosmos/relayer/v2/relayer/chains/cosmos/stride"
+	"github.com/cosmos/relayer/v2/relayer/ethermint"
 	"github.com/cosmos/relayer/v2/relayer/provider"
 	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
@@ -213,8 +216,14 @@ func (cc *CosmosProvider) broadcastTx(
 		cc.LogFailedTx(rlyResp, err, msgs)
 		return err
 	}
-
-	cc.UpdateFeesSpent(cc.ChainId(), cc.Key(), fees)
+	address, err := cc.Address()
+	if err != nil {
+		cc.log.Error(
+			"failed to get relayer bech32 wallet addresss",
+			zap.Error(err),
+		)
+	}
+	cc.UpdateFeesSpent(cc.ChainId(), cc.Key(), address, fees)
 
 	// TODO: maybe we need to check if the node has tx indexing enabled?
 	// if not, we need to find a new way to block until inclusion in a block
@@ -1218,7 +1227,29 @@ func (cc *CosmosProvider) queryTMClientState(ctx context.Context, srch int64, sr
 	clientState, ok := clientStateExported.(*tmclient.ClientState)
 	if !ok {
 		return &tmclient.ClientState{},
-			fmt.Errorf("error when casting exported clientstate to tendermint type")
+			fmt.Errorf("error when casting exported clientstate to tendermint type, got(%T)", clientStateExported)
+	}
+
+	return clientState, nil
+}
+
+// queryLocalhostClientState retrieves the latest consensus state for a client in state at a given height
+// and unpacks/cast it to localhost client state.
+func (cc *CosmosProvider) queryLocalhostClientState(ctx context.Context, srch int64) (*localhost.ClientState, error) {
+	clientStateRes, err := cc.QueryClientStateResponse(ctx, srch, ibcexported.LocalhostClientID)
+	if err != nil {
+		return &localhost.ClientState{}, err
+	}
+
+	clientStateExported, err := clienttypes.UnpackClientState(clientStateRes.ClientState)
+	if err != nil {
+		return &localhost.ClientState{}, err
+	}
+
+	clientState, ok := clientStateExported.(*localhost.ClientState)
+	if !ok {
+		return &localhost.ClientState{},
+			fmt.Errorf("error when casting exported clientstate to localhost client type, got(%T)", clientStateExported)
 	}
 
 	return clientState, nil
@@ -1257,7 +1288,7 @@ func (cc *CosmosProvider) NewClientState(
 	}, nil
 }
 
-func (cc *CosmosProvider) UpdateFeesSpent(chain, key string, fees sdk.Coins) {
+func (cc *CosmosProvider) UpdateFeesSpent(chain, key, address string, fees sdk.Coins) {
 	// Don't set the metrics in testing
 	if cc.metrics == nil {
 		return
@@ -1270,7 +1301,7 @@ func (cc *CosmosProvider) UpdateFeesSpent(chain, key string, fees sdk.Coins) {
 	for _, fee := range cc.TotalFees {
 		// Convert to a big float to get a float64 for metrics
 		f, _ := big.NewFloat(0.0).SetInt(fee.Amount.BigInt()).Float64()
-		cc.metrics.SetFeesSpent(chain, key, fee.GetDenom(), f)
+		cc.metrics.SetFeesSpent(chain, key, address, fee.GetDenom(), f)
 	}
 }
 
@@ -1340,6 +1371,10 @@ func (cc *CosmosProvider) PrepareFactory(txf tx.Factory) (tx.Factory, error) {
 		txf = txf.WithGas(cc.PCfg.MinGasAmount)
 	}
 
+	txf, err = cc.SetWithExtensionOptions(txf)
+	if err != nil {
+		return tx.Factory{}, err
+	}
 	return txf, nil
 }
 
@@ -1360,6 +1395,35 @@ func (cc *CosmosProvider) AdjustEstimatedGas(gasUsed uint64) (uint64, error) {
 		gas = math.Min(gas, float64(cc.PCfg.MaxGasAmount))
 	}
 	return uint64(gas), nil
+}
+
+// SetWithExtensionOptions sets the dynamic fee extension options on the given
+// transaction factory using the configuration options from the CosmosProvider.
+// The function creates an extension option for each configuration option and
+// serializes it into a byte slice before adding it to the list of extension
+// options. The function returns the updated transaction factory with the new
+// extension options or an error if the serialization fails or an invalid option
+// value is encountered.
+func (cc *CosmosProvider) SetWithExtensionOptions(txf tx.Factory) (tx.Factory, error) {
+	extOpts := make([]*types.Any, 0, len(cc.PCfg.ExtensionOptions))
+	for _, opt := range cc.PCfg.ExtensionOptions {
+		max, ok := sdk.NewIntFromString(opt.Value)
+		if !ok {
+			return txf, fmt.Errorf("invalid opt value")
+		}
+		extensionOption := ethermint.ExtensionOptionDynamicFeeTx{
+			MaxPriorityPrice: max,
+		}
+		extBytes, err := extensionOption.Marshal()
+		if err != nil {
+			return txf, err
+		}
+		extOpts = append(extOpts, &types.Any{
+			TypeUrl: "/ethermint.types.v1.ExtensionOptionDynamicFeeTx",
+			Value:   extBytes,
+		})
+	}
+	return txf.WithExtensionOptions(extOpts...), nil
 }
 
 // CalculateGas simulates a tx to generate the appropriate gas settings before broadcasting a tx.