+++ relayer/codecs/ethermint/codec.go
@@ -5,6 +5,7 @@ import (
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	"github.com/cosmos/cosmos-sdk/types/tx"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	"github.com/cosmos/relayer/v2/relayer/ethermint"
 )
 
 // RegisterInterfaces register the Ethermint key concrete types.
@@ -21,8 +22,8 @@ func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
 		&EthAccount{},
 	)
 	registry.RegisterImplementations(
-		(*tx.ExtensionOptionI)(nil),
+		(*tx.TxExtensionOptionI)(nil),
 		&ExtensionOptionsWeb3Tx{},
-		&ExtensionOptionDynamicFeeTx{},
+		&ethermint.ExtensionOptionDynamicFeeTx{},
 	)
 }