+++ relayer/processor/path_processor_internal.go
@@ -499,6 +499,7 @@ func (pp *PathProcessor) unrelayedChannelHandshakeMessages(
 			eventType: chantypes.EventTypeChannelOpenTry,
 			info:      info,
 		}
+
 		if pathEndChannelHandshakeMessages.Dst.shouldSendChannelMessage(
 			msgOpenTry, pathEndChannelHandshakeMessages.Src,
 		) {
@@ -1005,11 +1006,11 @@ func (pp *PathProcessor) processLatestMessages(ctx context.Context, cancel func(
 	// if sending messages fails to one pathEnd, we don't need to halt sending to the other pathEnd.
 	var eg errgroup.Group
 	eg.Go(func() error {
-		mp := newMessageProcessor(pp.log, pp.metrics, pp.memo, pp.clientUpdateThresholdTime)
+		mp := newMessageProcessor(pp.log, pp.metrics, pp.memo, pp.clientUpdateThresholdTime, pp.isLocalhost)
 		return mp.processMessages(ctx, pathEnd1Messages, pp.pathEnd2, pp.pathEnd1)
 	})
 	eg.Go(func() error {
-		mp := newMessageProcessor(pp.log, pp.metrics, pp.memo, pp.clientUpdateThresholdTime)
+		mp := newMessageProcessor(pp.log, pp.metrics, pp.memo, pp.clientUpdateThresholdTime, pp.isLocalhost)
 		return mp.processMessages(ctx, pathEnd2Messages, pp.pathEnd1, pp.pathEnd2)
 	})
 	return eg.Wait()